+++ feature/filters/src/main/java/com/t8rin/imagetoolbox/feature/filters/data/model/FastBilaterialBlurFilter.kt
@@ -26,7 +26,7 @@ import com.t8rin.imagetoolbox.core.domain.utils.roundTo
 import com.t8rin.imagetoolbox.core.filters.domain.model.Filter
 
 internal class FastBilaterialBlurFilter(
-    override val value: Triple<Int, Float, Float> = Triple(11, 10f, 3f),
+    override val value: Triple<Float, Float, Float> = Triple(11f, 10f, 3f),
 ) : Transformation<Bitmap>, Filter.FastBilaterialBlur {
 
     override val cacheKey: String
@@ -39,7 +39,7 @@ internal class FastBilaterialBlurFilter(
         bitmap = input,
         spatialSigma = value.second,
         rangeSigma = value.third,
-        kernelSize = value.first.toFloat().roundTo(NEAREST_ODD_ROUNDING).toInt()
+        kernelSize = value.first.roundTo(NEAREST_ODD_ROUNDING).toInt()
     )
 
 }
